Update models and tests to use BigDecimal for pricing and voltage

+++ src/main/java/com/martinbg/electricityapi/config/ElectricityConfig.java
@@ -3,48 +3,50 @@ package com.martinbg.electricityapi.config;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class ElectricityConfig {
 
     @Value("${electricity_day_price}")
-    private double dayPrice;
+    private BigDecimal dayPrice;
 
     @Value("${electricity_night_price}")
-    private double nightPrice;
+    private BigDecimal nightPrice;
 
     @Value("${electricity_low_voltage_transmission}")
-    private double lowVoltageTransmission;
+    private BigDecimal lowVoltageTransmission;
 
     @Value("${electricity_high_voltage_transmission}")
-    private double highVoltageTransmission;
+    private BigDecimal highVoltageTransmission;
 
     @Value("${electricity_low_voltage_access}")
-    private double lowVoltageAccess;
+    private BigDecimal lowVoltageAccess;
 
     @Value("${electricity_high_voltage_access}")
-    private double highVoltageAccess;
+    private BigDecimal highVoltageAccess;
 
-    public double getDayPrice() {
+    public BigDecimal getDayPrice() {
         return dayPrice;
     }
 
-    public double getNightPrice() {
+    public BigDecimal getNightPrice() {
         return nightPrice;
     }
 
-    public double getLowVoltageTransmission() {
+    public BigDecimal getLowVoltageTransmission() {
         return lowVoltageTransmission;
     }
 
-    public double getHighVoltageTransmission() {
+    public BigDecimal getHighVoltageTransmission() {
         return highVoltageTransmission;
     }
 
-    public double getLowVoltageAccess() {
+    public BigDecimal getLowVoltageAccess() {
         return lowVoltageAccess;
     }
 
-    public double getHighVoltageAccess() {
+    public BigDecimal getHighVoltageAccess() {
         return highVoltageAccess;
     }
