+++ src/test/java/com/martinbg/electricityapi/model/ElectricityModelTest.java
@@ -1,22 +1,25 @@
 package com.martinbg.electricityapi.model;
 
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+
+import java.math.BigDecimal;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ElectricityModelTest {
 
     @Test
     void testElectricityModelFields() {
-        ElectricityPrice electricityPrice = new ElectricityPrice(0.20, 0.10);
-        ElectricityVoltage lowVoltage = new ElectricityVoltage(0.01, 0.05);
-        ElectricityVoltage highVoltage = new ElectricityVoltage(0.02, 0.10);
+        ElectricityPrice electricityPrice = new ElectricityPrice(BigDecimal.valueOf(0.2), BigDecimal.valueOf(0.1));
+        ElectricityVoltage lowVoltage = new ElectricityVoltage(BigDecimal.valueOf(0.01), BigDecimal.valueOf(0.05));
+        ElectricityVoltage highVoltage = new ElectricityVoltage(BigDecimal.valueOf(0.02), BigDecimal.valueOf(0.1));
         ElectricityModel electricityModel = new ElectricityModel(electricityPrice, lowVoltage, highVoltage);
 
-        assertEquals(0.20, electricityModel.price().getDaily());
-        assertEquals(0.10, electricityModel.price().getNight());
-        assertEquals(0.01, electricityModel.lowVoltage().getAccess());
-        assertEquals(0.05, electricityModel.lowVoltage().getTransmission());
-        assertEquals(0.02, electricityModel.highVoltage().getAccess());
-        assertEquals(0.10, electricityModel.highVoltage().getTransmission());
+        assertEquals(BigDecimal.valueOf(0.20), electricityModel.price().getDaily());
+        assertEquals(BigDecimal.valueOf(0.10), electricityModel.price().getNight());
+        assertEquals(BigDecimal.valueOf(0.01), electricityModel.lowVoltage().getAccess());
+        assertEquals(BigDecimal.valueOf(0.05), electricityModel.lowVoltage().getTransmission());
+        assertEquals(BigDecimal.valueOf(0.02), electricityModel.highVoltage().getAccess());
+        assertEquals(BigDecimal.valueOf(0.10), electricityModel.highVoltage().getTransmission());
     }
 }
