+++ src/test/java/com/martinbg/electricityapi/model/ElectricityVoltageTest.java
@@ -2,14 +2,16 @@ package com.martinbg.electricityapi.model;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ElectricityVoltageTest {
 
     @Test
     void testLowVoltageConstructor() {
-        double access = 0.00977;
-        double transmission = 0.04660;
+        BigDecimal access = BigDecimal.valueOf(0.00977);
+        BigDecimal transmission = BigDecimal.valueOf(0.04660);
 
         ElectricityVoltage voltage = new ElectricityVoltage(access, transmission);
 
@@ -19,9 +21,9 @@ class ElectricityVoltageTest {
 
     @Test
     void testSetters() {
-        ElectricityVoltage voltage = new ElectricityVoltage(0.01, 0.05);
-        double newTransmission = 0.02;
-        double newAccess = 0.06;
+        ElectricityVoltage voltage = new ElectricityVoltage(BigDecimal.valueOf(0.01), BigDecimal.valueOf(0.05));
+        BigDecimal newTransmission = BigDecimal.valueOf(0.02);
+        BigDecimal newAccess = BigDecimal.valueOf(0.06);
 
         voltage.setAccess(newAccess);
         voltage.setTransmission(newTransmission);
